Create MovieTicket.java
 The program prompts the user to enter their age and the time of the movie.
Based on this input, it determines the price of the movie ticket.
The conditions are as follows:
Children under 5 and senior citizens (age 60 and above) get a free ticket. Matinee price is 250 for shows before noon.
Evening price is 300 for shows between 5 PM and 9 PM. Regular price is 200 for all other times.